feature: 开发课程的增删改，以及通用分页查询接口；优化了一些代码

+++ src/main/java/com/ujs/trainingprogram/tp/dao/entity/CourseDO.java
@@ -105,7 +105,7 @@ public class CourseDO extends BaseDO {
     /**
      * 选修学分要求
      */
-    private byte requiredElective;
+    private Integer requiredElective;
 
     /**
      * 建议修读学期
